requirements.txt
django
numpy==1.19.5
djangorestframework